Enable Less Expensive WellMatcher Construction
This commit switches the internal representation of WellMatcher from
NameOrder and WListManager objects to reference semantics.  This
enables using such objects by reference instead of copying the
intermediate size objects into internal storage every time we form a
WellMatcher object--for instance on every call to member function

    Schedule::wellNames()

We still allow value semantics for the NameOrder member, but this
now requires explicit transfer of ownership and the object will be
stored in a unique_ptr<>.

The motivation for this change is that WellMatcher objects are
usually rather short-lived, but we often create many such objects
during Schedule construction, summary evaluation, or UDQ operations.
The downside to doing this is that user code must be more careful
about object lifetimes.