Don't Move From Object Twice
We've had a long-standing problem in our linearisation logic for
restart file well data.  We were calling forward<>(obj) repeatedly
and the only reason this didn't blow up is that it captured a unit
system by reference.  Use a guaranteed safe call instead.

While here, also fix a similar problem for the group data and split
some long lines in preparation of altering the implementation of
ACTIONX match sets.