fixed: out-of-bounds indexing in FieldProps
this happens when grid processing removed cells (pinch, minpv, ..).
in particular the actnum in the grid_ptr member (ie EclipseState::getInputGrid())
does not reflect the additional inactive cells, and thus the
active_index in the Box are not the same as those in the FieldData
(which has been modified in the call to FieldProps::reset_actnum).
this causes oob indexing in handle_schedule_keywords / multiply_deck.

this is a minimal fix to fix the oob indexing. there might be more
dragons hidden in the output where the inconsistent actnum from the
input grid is used.