Generalise Entity Matching for UDQ Assignment
This commit introduces more general entity matching for UDQ
assignments.  This is in preparation of adding support for
assignment statements within an ACTIONX block using the "?" selector
to identify wells and/or groups.

We add new member functions

    UDQSet UDQAssign::eval(entities, matcher) const

with 'entities' being a vector of either well/group names or a
vector of enumerated items such as the known segments of all MS
wells and 'matcher' being a call-back function that translates a
selector to a vector of specific entities.  Those are, in turn,
expected to be a subset of the input entities.  The result value is
sized according to the input entities and only those entities
identified by the 'matcher' will have a defined value.

We reimplement the UDQConfig's well and group assignment operations
in terms of these new UDQAssign::eval() functions and in so doing
simplify the interfaces of these.  They can now be fully implemented
in terms of a UDQContext object so we no longer need the Schedule or
report_step parameters.  We therefore remove these from
UDQConfig::eval().