Generalise Action Result Entity Set
In particular, retire the original 'WellSet' container in favour of
a set implementation based on a sorted vector.  The latter could, in
principle, be more broadly useful, but we add it in the ActionResult
implementation file to gain more experience.  While here, also
replace the overloaded operators

    Result::operator bool()
    Result::operator|=()
    Result::operator&=()

with named member functions

    Result::conditionSatisfied()
    Result::makeSetUnion()
    Result::makeSetIntersection()

to (hopefully) make the code more self explanatory.  We furthermore
expose the set of matching entities as (custom) ranges instead of
vector<> objects in order to limit data copying.  Those ranges can
be linearised into vector<> objects if/when needed, however.

Make callers aware of the revised API for matching entities.  As in
the current master sources, we presently support wells only, but
expect to introduce group names and possibly others as well in the
future.