Support Dynamic Well Names in UDQ Assignments in ACTIONX Blocks
This commit adds support for using the special name selector '?' in
well level UDQ assignments within an ACTIONX block.  In particular,
we now support ACTIONX blocks like

    ACTIONX
    TESTW 25 1 /
    DAY > 10 AND /
    WMCTL '*PROD?' = 0 /
    /

    UDQ
     ASSIGN WU_TEST '?' 1 /
    /

    ENDACTIO

in which the selector '?' represents all wells from all well lists
matching the well list template '*PROD?' for which the WMCTL is
zero, meaning the well is shut.

To this end, make a tiny wrapper inside UDQConfig for

    Action::Result::MatchingEntities

that provides a way to forward those entities into

    UDQConfig::add_record()

without making UDQConfig itself depend on Action::Result.  We then
add a new private member function

    UDQConfig::add_assign_wellvar()

that knows how to forward dynamically selected well names onto the
pertinent UDQAssign records.  When affecting the UDQ assignment, we
furthermore account for the possibility that an assignment record
might have more than one entity in this case and use the 'pattern'
directly as a list of entities if its size is greater than one.
Otherwise, we continue to translate the 'pattern' through the well
name matcher.  Doing so is redundant if a '?' selector matched
exactly one well and we will monitor this choice for any performance
regressions.